backend/requirements.txt
websockets
python-dotenv
openai==1.69.0
fastapi[standard]
sqlmodel
redis[hiredis]